+++ src/vs/editor/browser/controller/hiddenArea/hiddenAreaState.ts
