+++ src/vs/editor/browser/controller/hiddenArea/hiddenWrapper.css
