+++ src/vs/editor/test/browser/controller/textAreaInput.test.ts
@@ -9,8 +9,8 @@ import { Emitter, Event } from 'vs/base/common/event';
 import { Disposable, DisposableStore } from 'vs/base/common/lifecycle';
 import { OperatingSystem } from 'vs/base/common/platform';
 import { ensureNoDisposablesAreLeakedInTestSuite } from 'vs/base/test/common/utils';
-import { ClipboardDataToCopy, IBrowser, ICompleteHiddenAreaWrapper, IHiddenAreaInputHost, HiddenAreaInput } from 'vs/editor/browser/controller/editContext/editContextInput';
-import { HiddenAreaState } from 'vs/editor/browser/controller/editContext/editContextState';
+import { ClipboardDataToCopy, IBrowser, ICompleteHiddenAreaWrapper, IHiddenAreaInputHost, HiddenAreaInput } from 'vs/editor/browser/controller/editContext/hiddenAreaInput';
+import { HiddenAreaState } from 'vs/editor/browser/controller/editContext/hiddenAreaState';
 import { Position } from 'vs/editor/common/core/position';
 import { IRecorded, IRecordedEvent, IRecordedTextareaState } from 'vs/editor/test/browser/controller/imeRecordedTypes';
 import { TestAccessibilityService } from 'vs/platform/accessibility/test/common/testAccessibilityService';
