+++ src/vs/editor/test/browser/controller/textAreaState.test.ts
@@ -6,7 +6,7 @@
 import assert from 'assert';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { ensureNoDisposablesAreLeakedInTestSuite } from 'vs/base/test/common/utils';
-import { HiddenAreaState, IHiddenAreaWrapper, PagedScreenReaderStrategy } from 'vs/editor/browser/controller/editContext/editContextState';
+import { HiddenAreaState, IHiddenAreaWrapper, PagedScreenReaderStrategy } from 'vs/editor/browser/controller/editContext/hiddenAreaState';
 import { Range } from 'vs/editor/common/core/range';
 import { Selection } from 'vs/editor/common/core/selection';
 import { createTextModel } from 'vs/editor/test/common/testTextModel';
